+++ .eslintrc.js
@@ -1,0 +1,13 @@
+module.exports = {
+    env: {
+        node: true,
+        es6: true
+    },
+    extends: 'eslint:recommended',
+    parserOptions: {
+        ecmaVersion: 2020
+    },
+    rules: {
+        // Add any specific rules here
+    }
+};